Python-snow-slack/requirements.txt
colorama==0.4.3
diffios==0.0.9
Jinja2==2.11.2
netmiko==3.3.2
python-dotenv==0.15.0
slack-sdk==3.4.0
slackclient==2.9.3
termcolor==1.1.0

